updates to 04

+++ examples/src/main/java/mixins/StatefulEscalate2.java
@@ -6,7 +6,7 @@ public interface StatefulEscalate2 extends Stateful {
 	String text();
 
 	default String escalated() {
-		int n = (Integer) getState(StatefulEscalate2.class, 0);
+		int n = getState(StatefulEscalate2.class, 0);
 		setState(StatefulEscalate2.class, n+1);
 
 		// generate n bangs, or empty strings for n=0
